Check existence of waitfile in test/integrationtest.py * test/integrationtest.py: when trying to delete a stale waitfile,   check for its existence first

+++ test/integrationtest.py
@@ -103,7 +103,9 @@ class IntegrationTest(unittest.TestCase):
         def testFunc(self):
             # Start listener and informer processes
             waitFile = 'test/%s-listener-ready' % listenerLang
-            os.remove(waitFile)
+            if os.exists(waitFile):
+                self.__logger.warn("Deleting old waitFile %s" % waitFile)
+                os.remove(waitFile)
             listenerProc = self.startProcess(listenerLang, "listener")
             waitStart = time.time()
             while not os.path.exists(waitFile):
